LayoutTests/sputnik/Conformance/12_Statement/12.5_The_if_Statement/S12.5_A6_T1-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected keyword 'true'. Expected '(' to start an 'if' condition.
S12.5_A6_T1

PASS Expected parsing failure

TEST COMPLETE
